Move test fixture to own file
